Moved the What IS Open Sub View Models to their own file. Set up the WhatIsOpenViewModelType protocol, but it has to be filled in before we make the views use the protocol type for the type of their vm property. Added logic to the WhatIsOpenPlace so that when a place is closed but will open next tomorrow it says that it'll open tomorrow instead of a date
