+++ UOKiosk/WhatIsOpen/WhatIsOpenSubViews.swift
